Add the elementary SequenceAligner root class
SequenceAligner is essentially designed to abstract away the existence of
Automatrix from alignment implementations. Ideally, clients will not need to
be aware of anything Automatrix itself does (they are, however, advised to
execute their own traversal of the matrix to infill – the implementation of
backfilling is pretty naïve. In the longer run, a population strategy might be
generated for this.